+++ src/Service/UserBlocker.php
@@ -4,6 +4,7 @@ namespace Mediawiki\Api\Service;
 
 use InvalidArgumentException;
 use Mediawiki\Api\MediawikiApi;
+use Mediawiki\Api\Options\BlockOptions;
 use Mediawiki\DataModel\User;
 
 /**
@@ -17,20 +18,6 @@ class UserBlocker {
 	private $api;
 
 	/**
-	 * @var array of params allowed in a block calls $options parameter
-	 */
-	private $allowedBlockOptions = array(
-		'anononly', // Block anonymous users only (i.e. disable anonymous edits for this IP)
-		'nocreate', // Prevent account creation
-		'autoblock', // Automatically block the last used IP address, and any subsequent IP addresses they try to login from
-		'noemail', // Prevent user from sending email through the wiki. (Requires the "blockemail" right.)
-		'hidename', // Hide the username from the block log. (Requires the "hideuser" right.)
-		'allowusertalk', // Allow the user to edit their own talk page (depends on $wgBlockAllowsUTEdit)
-		'reblock', // If the user is already blocked, overwrite the existing block
-		'watchuser', // Watch the user/IP's user and talk pages
-	);
-
-	/**
 	 * @param MediawikiApi $api
 	 */
 	public function __construct( MediawikiApi $api ) {
@@ -41,14 +28,12 @@ class UserBlocker {
 	 * @since 0.3
 	 *
 	 * @param User|string $user
-	 * @param string $expiry
-	 * @param string $reason
-	 * @param string[] $options @see $this->allowedBlockOptions
+	 * @param BlockOptions $options
 	 *
 	 * @throws InvalidArgumentException
 	 * @return bool
 	 */
-	public function block( $user, $expiry = 'never', $reason = '', $options = array() ) {
+	public function block( $user, BlockOptions $options = null ) {
 		if( !$user instanceof User && !is_string( $user ) ) {
 			throw new InvalidArgumentException( '$user must be either a string or User object' );
 		}
@@ -61,16 +46,36 @@ class UserBlocker {
 			'user' => $user,
 			'token' => $this->api->getToken( 'block' ),
 		);
-		if( $expiry !== 'never' ) {
-			$params['expiry'] = $expiry;
+		if( $options->getExpiry() !== 'never' ) {
+			$params['expiry'] = $options->getExpiry();
 		}
+		$reason = $options->getReason();
 		if( !empty( $reason ) ) {
 			$params['reason'] = $reason;
 		}
-		foreach( $options as $option ) {
-			if( in_array( $option, $this->$allowedBlockOptions ) ) {
-				$params[$option] = '';
-			}
+		if( $options->getAllowusertalk() ){
+			$params['allowusertalk'] = '';
+		}
+		if( $options->getAnononly() ){
+			$params['allowusertalk'] = '';
+		}
+		if( $options->getAutoblock() ){
+			$params['allowusertalk'] = '';
+		}
+		if( $options->getHidename() ){
+			$params['allowusertalk'] = '';
+		}
+		if( $options->getNocreate() ){
+			$params['allowusertalk'] = '';
+		}
+		if( $options->getNoemail() ){
+			$params['allowusertalk'] = '';
+		}
+		if( $options->getReblock() ){
+			$params['allowusertalk'] = '';
+		}
+		if( $options->getWatchuser() ){
+			$params['allowusertalk'] = '';
 		}
 
 		$this->api->postAction( 'block', $params );